# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ALU Marketplace API",
    "version": "1.0.0",
    "description": "API documentation for ALU Marketplace"
  },
  "basePath": "/api",
  "paths": {
    "/products": {
      "get": {
        "summary": "Get all products",
        "responses": {
          "200": {
            "description": "A list of products"
          }
        }
      },
      "post": {
        "summary": "Create new product",
        "description": "Add a new product to the marketplace (admin only)",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "ALU Branded Notebook"
                },
                "description": {
                  "type": "string",
                  "example": "High-quality hardcover notebook with ALU logo"
                },
                "price": {
                  "type": "number",
                  "format": "float",
                  "example": 12.99
                },
                "categoryId": {
                  "type": "integer",
                  "example": 3
                },
                "stockQuantity": {
                  "type": "integer",
                  "example": 100
                },
                "images": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "example": "https://storage.alu-marketplace.com/products/notebook-blue.jpg"
                  }
                },
                "features": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "example": "100 pages"
                  }
                }
              },
              "required": ["name", "description", "price", "categoryId", "stockQuantity"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Product created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "example": 42
                },
                "name": {
                  "type": "string",
                  "example": "ALU Branded Notebook"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2025-03-31T10:15:30Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      }
    },
    "/product-categories": {
      "get": {
        "summary": "Get all product categories",
        "responses": {
          "200": {
            "description": "A list of product categories"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Get all users",
        "description": "Retrieve a list of all users (admin only)",
        "responses": {
          "200": {
            "description": "A list of users",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 1
                  },
                  "username": {
                    "type": "string",
                    "example": "johndoe"
                  },
                  "email": {
                    "type": "string",
                    "example": "john.doe@alu.edu"
                  },
                  "role": {
                    "type": "string",
                    "example": "student"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      },
      "post": {
        "summary": "Create user",
        "description": "Create a new user (admin only)",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "newuser"
                },
                "email": {
                  "type": "string",
                  "example": "new.user@alu.edu"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "example": "securepassword"
                },
                "role": {
                  "type": "string",
                  "example": "student"
                }
              },
              "required": ["username", "email", "password", "role"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Bad request - Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Register new user",
        "description": "Register a new user account",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "newstudent"
                },
                "email": {
                  "type": "string",
                  "example": "student@alu.edu"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "example": "securepassword"
                },
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                }
              },
              "required": ["username", "email", "password", "firstName", "lastName"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Registration successful",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "example": 42
                },
                "username": {
                  "type": "string",
                  "example": "newstudent"
                },
                "message": {
                  "type": "string",
                  "example": "Registration successful"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input or email already registered"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User login",
        "description": "Authenticate user and return a token",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "testuser"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "example": "securepassword"
                }
              },
              "required": ["username", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials"
          }
        }
      }
    },
    "/payments": {
      "get": {
        "summary": "Get payment history",
        "description": "Get payment history for the authenticated user",
        "responses": {
          "200": {
            "description": "A list of payments",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 1
                  },
                  "amount": {
                    "type": "number",
                    "format": "float",
                    "example": 150.75
                  },
                  "status": {
                    "type": "string",
                    "example": "completed"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "example": "credit_card"
                  },
                  "createdAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-03-15T14:30:00Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create payment",
        "description": "Process a new payment",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "orderId": {
                  "type": "integer",
                  "example": 42
                },
                "amount": {
                  "type": "number",
                  "format": "float",
                  "example": 150.75
                },
                "paymentMethod": {
                  "type": "string",
                  "example": "credit_card"
                },
                "cardDetails": {
                  "type": "object",
                  "properties": {
                    "cardNumber": {
                      "type": "string",
                      "example": "4242424242424242"
                    },
                    "expiryMonth": {
                      "type": "integer",
                      "example": 12
                    },
                    "expiryYear": {
                      "type": "integer",
                      "example": 2028
                    },
                    "cvc": {
                      "type": "string",
                      "example": "123"
                    }
                  },
                  "required": ["cardNumber", "expiryMonth", "expiryYear", "cvc"]
                }
              },
              "required": ["orderId", "amount", "paymentMethod", "cardDetails"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Payment processed successfully",
            "schema": {
              "type": "object",
              "properties": {
                "paymentId": {
                  "type": "string",
                  "example": "pay_1Lmn5EJ9qvKYlo2CWWc5bOPH"
                },
                "status": {
                  "type": "string",
                  "example": "completed"
                },
                "receiptUrl": {
                  "type": "string",
                  "example": "https://pay.stripe.com/receipts/..."
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid payment details"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/cart": {
      "get": {
        "summary": "Get user cart",
        "description": "Get the current user's shopping cart",
        "responses": {
          "200": {
            "description": "Shopping cart details",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "example": 15
                },
                "items": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "productId": {
                        "type": "integer",
                        "example": 7
                      },
                      "name": {
                        "type": "string",
                        "example": "ALU Hoodie"
                      },
                      "quantity": {
                        "type": "integer",
                        "example": 2
                      },
                      "price": {
                        "type": "number",
                        "format": "float",
                        "example": 45.99
                      },
                      "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 91.98
                      }
                    }
                  }
                },
                "totalItems": {
                  "type": "integer",
                  "example": 3
                },
                "totalAmount": {
                  "type": "number",
                  "format": "float",
                  "example": 124.97
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Add item to cart",
        "description": "Add a product to the user's shopping cart",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": "integer",
                  "example": 7
                },
                "quantity": {
                  "type": "integer",
                  "example": 1
                }
              },
              "required": ["productId", "quantity"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item added to cart successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Item added to cart successfully"
                },
                "cartId": {
                  "type": "integer",
                  "example": 15
                },
                "totalItems": {
                  "type": "integer",
                  "example": 3
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid product ID or quantity"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/orders": {
      "get": {
        "summary": "Get user orders",
        "description": "Get order history for the authenticated user",
        "responses": {
          "200": {
            "description": "A list of orders",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 24
                  },
                  "orderNumber": {
                    "type": "string",
                    "example": "ALU-ORD-2025-0042"
                  },
                  "status": {
                    "type": "string",
                    "example": "delivered"
                  },
                  "orderDate": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-03-10T09:15:00Z"
                  },
                  "totalAmount": {
                    "type": "number",
                    "format": "float",
                    "example": 124.97
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "integer",
                          "example": 7
                        },
                        "name": {
                          "type": "string",
                          "example": "ALU Hoodie"
                        },
                        "quantity": {
                          "type": "integer",
                          "example": 2
                        },
                        "price": {
                          "type": "number",
                          "format": "float",
                          "example": 45.99
                        }
                      }
                    }
                  },
                  "shippingAddress": {
                    "type": "object",
                    "properties": {
                      "street": {
                        "type": "string",
                        "example": "123 Main St"
                      },
                      "city": {
                        "type": "string",
                        "example": "Kigali"
                      },
                      "country": {
                        "type": "string",
                        "example": "Rwanda"
                      },
                      "postalCode": {
                        "type": "string",
                        "example": "00000"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}