# routes/cart.js
const express = require('express');
const router = express.Router();
const auth = require('../middleware/auth');

// Placeholder controller
const cartController = {
  getCart: (req, res) => {
    res.status(200).json({
      success: true,
      message: 'Cart retrieved successfully',
      data: { items: [], total: 0 }
    });
  },
  addToCart: (req, res) => {
    res.status(200).json({
      success: true,
      message: 'Item added to cart successfully',
      data: { items: [req.body], total: req.body.price }
    });
  },
  updateCartItem: (req, res) => {
    res.status(200).json({
      success: true,
      message: 'Cart item updated successfully',
      data: { items: [req.body], total: req.body.price }
    });
  },
  removeFromCart: (req, res) => {
    res.status(200).json({
      success: true,
      message: 'Item removed from cart successfully',
      data: { items: [], total: 0 }
    });
  }
};

router.get('/', auth.verifyToken, cartController.getCart);
router.post('/add', auth.verifyToken, cartController.addToCart);
router.put('/:itemId', auth.verifyToken, cartController.updateCartItem);
router.delete('/:itemId', auth.verifyToken, cartController.removeFromCart);

module.exports = router;